Add AbstractBiomedisaSegmentEditorEffect and add the 3D preview to Biomedisa
